refactor: Remove redundant headers
