+++ frontend/src/features/cart/components/index.ts
@@ -1,5 +1,3 @@
 export { default as CartEmptyList } from './CartEmptyList';
-export { default as CartHeader } from './CartHeader';
 export { default as CartList } from './CartList';
 export { default as CartSummary } from './CartSummary';
-
