+++ frontend/src/features/cart/pages/CartPage.tsx
@@ -1,7 +1,6 @@
 import { useState } from "react";
 import Button from "../../../components/Button";
 import CartEmptyList from "../components/CartEmptyList";
-import CartHeader from "../components/CartHeader";
 import CartList from "../components/CartList";
 import CartSummary from "../components/CartSummary";
 import { cartData } from "../components/mockData";
@@ -29,34 +28,33 @@ const CartPage = () => {
     }
     return (
         <div className="min-h-screen bg-gray-50">
-            <header className="bg-white border-b border-gray-200 sticky top-0 z-10">
-                <CartHeader
-                    isEditing={isEditing}
-                    onToggleEdit={() => setIsEditing(!isEditing)}
-                />
-            </header>
-            <main>
-                <CartList
-                    items={cartItems}
-                    isEditing={isEditing}
-                    onRemoveItem={removeItem}
-                />
-                <div className="mx-4 my-6">
-                    <CartSummary
-                        totalAmount={totalAmount}
-                        shippingFee={shippingFee}
-                        finalAmount={finalAmount}
-                        formatPrice={formatPrice}
+            <div className="max-w-3xl mx-auto">
+                <main className="pb-24">
+                    <CartList
+                        items={cartItems}
+                        isEditing={isEditing}
+                        onToggleEdit={() => setIsEditing(!isEditing)}
+                        onRemoveItem={removeItem}
                     />
+                    <div className="mx-4 my-6">
+                        <CartSummary
+                            totalAmount={totalAmount}
+                            shippingFee={shippingFee}
+                            finalAmount={finalAmount}
+                            formatPrice={formatPrice}
+                        />
+                    </div>
+                </main>
+                <div className="fixed bottom-0 left-0 right-0 max-w-3xl mx-auto">
+                    <div className="p-4 bg-white">
+                        <Button
+                            onClick={() => { }}
+                            className="w-full bg-blue-600 text-white py-4 font-semibold text-lg cursor-pointer whitespace-nowrap hover:bg-blue-700 transition-colors rounded-lg"
+                        >
+                            결제하기 · {formatPrice(finalAmount)}
+                        </Button>
+                    </div>
                 </div>
-            </main>
-            <div className="fixed bottom-0 left-0 right-0 p-4">
-                <Button
-                    onClick={() => { }}
-                    className="w-full bg-blue-600 text-white py-4 font-semibold text-lg cursor-pointer whitespace-nowrap hover:bg-blue-700 transition-colors rounded-lg"
-                >
-                    결제하기 · {formatPrice(finalAmount)}
-                </Button>
             </div>
         </div>
     );
