--- frontend/src/features/cart/types/CartHeaderProps.ts
@@ -1,4 +1,0 @@
-export interface CartHeaderProps {
-    isEditing: boolean;
-    onToggleEdit: () => void;
-}
