+++ frontend/src/features/cart/types/CartListProps.ts
@@ -3,5 +3,6 @@ import type { Cart } from "./Cart";
 export interface CartListProps {
     items: Cart[];
     isEditing: boolean;
+    onToggleEdit: () => void;
     onRemoveItem: (id: number) => void;
 }
