+++ frontend/src/features/order/components/index.ts
@@ -1,3 +1,2 @@
-export { default as OrderHeader } from './OrderHeader';
 export { default as OrderSearch } from './OrderSearch';
 export { default as OrderList } from './OrderList';
