+++ frontend/src/features/product/components/auction/index.ts
@@ -1,5 +1,3 @@
-export { default as AuctionHeader } from './AuctionHeader';
 export { default as AuctionList } from './AuctionList';
 export { default as AuctionModal } from './AuctionModal';
 export { default as AuctionSearch } from './AuctionSearch';
-
