+++ frontend/src/features/product/components/brand/index.ts
@@ -1,4 +1,3 @@
-export { default as BrandHeader } from './BrandHeader';
 export { default as BrandList } from './BrandList';
-export { default as AuctionHeader } from './BrandModal';
+export { default as BrandModal } from './BrandModal';
 export { default as BrandPopular } from './BrandPopular';
