+++ frontend/src/features/product/types/brand/BrandModalProps.ts
@@ -4,5 +4,5 @@ export interface BrandModalProps {
     isOpen: boolean;
     onClose: () => void;
     product: BrandProduct | null;
-    onSubmit: (bidAmount: number) => void;
+    onSubmit: (productId: number, bidAmount: number) => void;
 }
