+++ frontend/src/features/profile/components/index.ts
@@ -1,4 +1,3 @@
-export { default as ProfileHeader } from './ProfileHeader';
 export { default as ProfileInfo } from './ProfileInfo';
 export { default as ProfileMenu } from './ProfileMenu';
 export { default as ProfileBid } from './ProfileBid';
