+++ frontend/src/features/profile/pages/ProfilePage.tsx
@@ -1,12 +1,9 @@
-import { ProfileAccount, ProfileBid, ProfileHeader, ProfileInfo, ProfileMenu, ProfileStats, ProfileWish } from "../components";
+import { ProfileAccount, ProfileBid, ProfileInfo, ProfileMenu, ProfileStats, ProfileWish } from "../components";
 import ProfilePoint from "../components/ProfilePoint";
 
 const ProfilePage = () => {
     return (
         <div className="min-h-screen bg-gray-50">
-            <header className="bg-white shadow-sm">
-                <ProfileHeader />
-            </header>
             <main className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-8">
                 <div className="grid grid-cols-12 gap-8">
                     {/* Left Sidebar */}
