+++ frontend/src/features/wish/components/index.ts
@@ -1,3 +1,2 @@
-export { default as WishHeader } from './WishHeader';
 export { default as WishList } from './WishList';
 export { default as WishSearch } from './WishSearch';
