--- frontend/src/features/wish/types/WishHeaderProps.ts
@@ -1,3 +1,0 @@
-export interface WishHeaderProps {
-    itemCount: number;
-}
